+++ extensions/menubar-calendar/src/types/preferences.ts
@@ -5,10 +5,12 @@ interface Preferences {
   menubarStyle: string;
   menubarIconStyle: string;
   menubarItemIconStyle: string;
+  showEventsInMenubar: string;
   largeCalendar: boolean;
   highlightCalendar: boolean;
   showWeekNumber: boolean;
   showCalendar: boolean;
+  dateFormat: string;
   calendarView: string;
   remindersView: string;
   showReminders: boolean;
@@ -20,9 +22,11 @@ export const {
   menubarStyle,
   menubarIconStyle,
   menubarItemIconStyle,
+  showEventsInMenubar,
   largeCalendar,
   highlightCalendar,
   showWeekNumber,
+  dateFormat,
   calendarView,
   remindersView,
   showCalendar,